pyproject.toml
[tool.poetry]
name = "utf-tech"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "utf"}]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
